Implement pass through Flow, rework decouple test assertions
